+++ modules/Material/ListItems/SimpleMenu.qml
@@ -1,23 +1,47 @@
 import QtQuick 2.0
 import Material 0.1
+import Material.Extras 0.1
 
 Subtitled {
     id: listItem
 
-    property alias model: listView.model
+    property var model
     property alias selectedIndex: listView.currentIndex
 
     subText: listView.currentItem.text
 
     onClicked: menu.open(listItem, units.dp(16), 0)
 
+    Label {
+        id: hiddenLabel
+        style: "subheading"
+        visible: false
+    }
+
+    property int columnWidth: {
+        var maxWidth = 0
+
+        for (var i = 0; i < model.length; i++) {
+            hiddenLabel.text = model[i]
+            
+            maxWidth = Math.max(maxWidth, hiddenLabel.implicitWidth + units.dp(33))
+        }
+
+        return maxWidth
+    }
+
     Dropdown {
         id: menu
 
         anchor: Item.TopLeft
 
-        width: Math.max(units.dp(56 * 2), Math.min(listItem.width - 2 * x, listView.contentWidth))
-        height: Math.min(10 * units.dp(48) + units.dp(16), listView.contentHeight + units.dp(16))
+        width: Math.max(units.dp(56 * 2), Math.min(listItem.width - units.dp(32), columnWidth))
+        height: Math.min(10 * units.dp(48) + units.dp(16), model.length * units.dp(48) + units.dp(16))
+
+        Rectangle {
+            anchors.fill: parent
+            radius: units.dp(2)
+        }
 
         ListView {
             id: listView
@@ -30,7 +54,8 @@ Subtitled {
             }
 
             interactive: false
-            height: count > 0 ? contentHeight : 0
+            height: contentHeight
+            model: listItem.model
 
             delegate: Standard {
                 id: delegateItem