+++ CHANGELOG.md
@@ -16,6 +16,12 @@ QML Material CHANGELOG
  * [Tests] Add UI tests for the ActionBar component
  * [IconButton] Add animation on mouse over
  * [ListItem] Add a new expandable section header list item
+ * [Dialog] Add support for full-width content
+ * [Icons] Update the FontAwesome parsing script to pull from the website
+ * [ListItem] Correctly calculate the width of a simple menu item
+ * [ActionBar] Hide non-visible actions in the action bar
+ * [Icon] Support custom icons with an iconSource property
+ * [ActionBar] Support disabling actions in the action bar
 
 
 ### Version 0.0.6 (April 16, 2015)